refactor: move gridStyles util to utils folder
